+++ docs/source/changelog.rst
@@ -5,6 +5,10 @@ Changelog
 
 **Upgrade Procedure:** Please follow :ref:`upgrading-kegbot` for general upgrade steps.
 
+Current Version (in development)
+--------------------------------
+* New beverage fields: IBU, SRM, star rating, and color.
+
 Version 1.0.2 (2014-08-21)
 --------------------------
 * Bugfix: Issue #309 (cannot reset password on private sites).